# util/interval: Remove Overlapper interface from IntervalTree