--- a/util/interval/interval.go
+++ b/util/interval/interval.go
@@ -44,27 +44,34 @@ func rangeError(r Range) error {
 	}
 }
 
-// An Overlapper can determine whether it overlaps a range.
-type Overlapper interface {
-	// Overlap returns a boolean indicating whether the receiver overlaps the parameter.
-	Overlap(Range) bool
-}
-
 // A Range is a type that describes the basic characteristics of an interval.
 type Range struct {
 	Start, End Comparable
 }
 
+// OverlapInclusive returns whether the two provided ranges overlap. It
+// defines overlapping as a pair of ranges that share a segment, with both
+// start and end keys treated as inclusive values.
+func (r Range) OverlapInclusive(other Range) bool {
+	return r.End.Compare(other.Start) >= 0 && r.Start.Compare(other.End) <= 0
+}
+
+// OverlapExclusive returns whether the two provided ranges overlap. It defines
+// overlapping as a pair of ranges that share a segment of the keyspace, with the
+// start keys treated as inclusive and the end keys treated as exclusive.
+func (r Range) OverlapExclusive(other Range) bool {
+	return r.End.Compare(other.Start) > 0 && r.Start.Compare(other.End) < 0
+}
+
 // An Interface is a type that can be inserted into a Tree.
 type Interface interface {
-	Overlapper
 	Range() Range
 	// Returns a unique ID for the element.
 	// TODO(nvanbenschoten) Should this be changed to an int64?
 	ID() uintptr
 }
 
-// A Comparable is a type that describes the ends of an Overlapper.
+// A Comparable is a type that describes the ends of a Range.
 type Comparable []byte
 
 // Compare returns a value indicating the sort order relationship between the
@@ -89,8 +96,9 @@ type Node struct {
 
 // A Tree manages the root node of an interval tree. Public methods are exposed through this type.
 type Tree struct {
-	Root  *Node // Root node of the tree.
-	Count int   // Number of elements stored.
+	Root       *Node                   // root node of the tree.
+	Count      int                     // number of elements stored.
+	Overlapper func(Range, Range) bool // determines how to define Range overlap.
 }
 
 // Helper methods
@@ -225,11 +233,10 @@ func (t *Tree) Len() int {
 	return t.Count
 }
 
-// Get returns a slice of Interfaces that overlap q in the Tree according
-// to q.Overlap().
-func (t *Tree) Get(q Overlapper) (o []Interface) {
-	if t.Root != nil && q.Overlap(t.Root.Range) {
-		t.Root.doMatch(func(e Interface) (done bool) { o = append(o, e); return }, q)
+// Get returns a slice of Interfaces that overlap r in the Tree.
+func (t *Tree) Get(r Range) (o []Interface) {
+	if t.Root != nil && t.Overlapper(r, t.Root.Range) {
+		t.Root.doMatch(func(e Interface) (done bool) { o = append(o, e); return }, r, t.Overlapper)
 	}
 	return
 }
@@ -395,7 +402,7 @@ func (t *Tree) Delete(e Interface, fast bool) (err error) {
 	if err := rangeError(r); err != nil {
 		return err
 	}
-	if t.Root == nil || !e.Overlap(t.Root.Range) {
+	if t.Root == nil || !t.Overlapper(r, t.Root.Range) {
 		return
 	}
 	var d int
@@ -612,66 +619,67 @@ func (n *Node) doReverse(fn Operation) (done bool) {
 	return
 }
 
-// DoMatching performs fn on all intervals stored in the tree that match q according to Overlap, with
-// q.Overlap() used to guide tree traversal, so DoMatching() will out perform Do() with a called
-// conditional function if the condition is based on sort order, but can not be reliably used if
-// the condition is independent of sort order. A boolean is returned indicating whether the Do
-// traversal was interrupted by an Operation returning true. If fn alters stored intervals' sort
-// relationships, future tree operation behaviors are undefined.
-func (t *Tree) DoMatching(fn Operation, q Overlapper) bool {
-	if t.Root != nil && q.Overlap(t.Root.Range) {
-		return t.Root.doMatch(fn, q)
+// DoMatching performs fn on all intervals stored in the tree that match r according to
+// t.Overlapper, with Overlapper() used to guide tree traversal, so DoMatching() will
+// outperform Do() with a called conditional function if the condition is based on sort
+// order, but can not be reliably used if the condition is independent of sort order. A
+// boolean is returned indicating whether the Do traversal was interrupted by an Operation
+// returning true. If fn alters stored intervals' sort relationships, future tree operation
+// behaviors are undefined.
+func (t *Tree) DoMatching(fn Operation, r Range) bool {
+	if t.Root != nil && t.Overlapper(r, t.Root.Range) {
+		return t.Root.doMatch(fn, r, t.Overlapper)
 	}
 	return false
 }
 
-func (n *Node) doMatch(fn Operation, q Overlapper) (done bool) {
-	if n.Left != nil && q.Overlap(n.Left.Range) {
-		done = n.Left.doMatch(fn, q)
+func (n *Node) doMatch(fn Operation, r Range, overlaps func(Range, Range) bool) (done bool) {
+	if n.Left != nil && overlaps(r, n.Left.Range) {
+		done = n.Left.doMatch(fn, r, overlaps)
 		if done {
 			return
 		}
 	}
-	if q.Overlap(n.Elem.Range()) {
+	if overlaps(r, n.Elem.Range()) {
 		done = fn(n.Elem)
 		if done {
 			return
 		}
 	}
-	if n.Right != nil && q.Overlap(n.Right.Range) {
-		done = n.Right.doMatch(fn, q)
+	if n.Right != nil && overlaps(r, n.Right.Range) {
+		done = n.Right.doMatch(fn, r, overlaps)
 	}
 	return
 }
 
-// DoMatchingReverse performs fn on all intervals stored in the tree that match q according to Overlap,
-// with q.Overlap() used to guide tree traversal, so DoMatching() will out perform Do() with a called
-// conditional function if the condition is based on sort order, but can not be reliably used if
-// the condition is independent of sort order. A boolean is returned indicating whether the Do
-// traversal was interrupted by an Operation returning true. If fn alters stored intervals' sort
-// relationships, future tree operation behaviors are undefined.
-func (t *Tree) DoMatchingReverse(fn Operation, q Overlapper) bool {
-	if t.Root != nil && q.Overlap(t.Root.Range) {
-		return t.Root.doMatch(fn, q)
+// DoMatchingReverse performs fn on all intervals stored in the tree that match r according to
+// t.Overlapper, with Overlapper() used to guide tree traversal, so DoMatching() will outperform
+// Do() with a called conditional function if the condition is based on sort order, but can not
+// be reliably used if the condition is independent of sort order. A boolean is returned indicating
+// whether the Do traversal was interrupted by an Operation returning true. If fn alters stored
+// intervals' sort relationships, future tree operation behaviors are undefined.
+func (t *Tree) DoMatchingReverse(fn Operation, r Range) bool {
+	if t.Root != nil && t.Overlapper(r, t.Root.Range) {
+		return t.Root.doMatch(fn, r, t.Overlapper)
 	}
 	return false
 }
 
-func (n *Node) doMatchReverse(fn Operation, q Overlapper) (done bool) {
-	if n.Right != nil && q.Overlap(n.Right.Range) {
-		done = n.Right.doMatchReverse(fn, q)
+func (n *Node) doMatchReverse(fn Operation, r Range, overlaps func(Range, Range) bool) (done bool) {
+	if n.Right != nil && overlaps(r, n.Right.Range) {
+		done = n.Right.doMatchReverse(fn, r, overlaps)
 		if done {
 			return
 		}
 	}
-	if q.Overlap(n.Elem.Range()) {
+	if overlaps(r, n.Elem.Range()) {
 		done = fn(n.Elem)
 		if done {
 			return
 		}
 	}
-	if n.Left != nil && q.Overlap(n.Left.Range) {
-		done = n.Left.doMatchReverse(fn, q)
+	if n.Left != nil && overlaps(r, n.Left.Range) {
+		done = n.Left.doMatchReverse(fn, r, overlaps)
 	}
 	return
 }