@tamird 2016-02-25
This is making me wonder: are we getting into the weeds with some of these
optimizations? 3% is nice, but this is a 100 row deletion, which I expect
to be rare in OLTP workloads.

On Wed, Feb 24, 2016 at 8:18 PM, Nathan VanBenschoten <
notifications@github.com> wrote:

> Related to #4616 <https://github.com/cockroachdb/cockroach/issues/4616>.
>
> Remove the Overlapper interface and directly compare Ranges when
> testing for overlapping. This reduces the genericity of the IntervalTree,
> but allows the interval tree to determine overlapping intervals without
> having to allocate. This is critical for both the TimestampCache and the
> RangeGroup.
>
> To support the two current use cases for IntervalTree, this change adds
> the notion of inclusive and exclusive overlapping. With inclusive
> overlapping, ranges' start and end keys are considered inclusive. With
> exclusive overlapping, ranges' start keys are considered inclusive, and
> end keys are considered exclusive.
>
> The PR also contains a secondary commit which makes sure RangeGroup's
> key variable in RangeGroup.Add only escapes when needed.
> Previously, RangeGroup.Add would allow its rangeKey to escape to the
> heap during every call, even though this was only needed during the
> case where the key needed to be inserted into the tree. By using a
> helper function to contain this insertion, we can avoid the variable
> from escaping for the vast majority of cases. This change gave a modest
> performance boost.
>
> name                              old time/op    new time/op    delta
> Bank_Cockroach-4                     267µs ± 6%     269µs ± 8%    ~    (p=0.661 n=10+9)
> Select1_Cockroach-4                 40.9µs ± 3%    40.6µs ± 1%    ~    (p=0.218 n=10+10)
> Select2_Cockroach-4                  755µs ± 3%     752µs ± 1%    ~    (p=0.780 n=10+9)
> Insert1_Cockroach-4                  400µs ± 1%     400µs ± 2%    ~    (p=0.796 n=9+9)
> Insert10_Cockroach-4                 688µs ± 1%     685µs ± 1%    ~    (p=0.077 n=9+9)
> Insert100_Cockroach-4               3.44ms ± 0%    3.39ms ± 1%  -1.44% (p=0.000 n=9+10)
> Update1_Cockroach-4                  661µs ± 1%     662µs ± 1%    ~    (p=0.739 n=10+10)
> Update10_Cockroach-4                1.28ms ± 3%    1.28ms ± 1%    ~    (p=0.905 n=10+9)
> Update100_Cockroach-4               6.83ms ± 0%    6.76ms ± 1%  -0.98% (p=0.000 n=8+8)
> Delete1_Cockroach-4                  617µs ± 2%     621µs ± 1%    ~    (p=0.243 n=9+10)
> Delete10_Cockroach-4                1.69ms ± 3%    1.69ms ± 3%    ~    (p=0.829 n=10+8)
> Delete100_Cockroach-4               13.1ms ± 4%    12.6ms ± 2%  -3.37% (p=0.006 n=10+8)
> Scan1_Cockroach-4                    213µs ± 2%     214µs ± 1%    ~    (p=0.079 n=10+9)
> Scan10_Cockroach-4                   253µs ± 1%     255µs ± 3%    ~    (p=0.278 n=9+10)
> Scan100_Cockroach-4                  563µs ± 2%     562µs ± 2%    ~    (p=1.000 n=9+10)
> PgbenchExec_Cockroach-4             3.12ms ± 4%    3.06ms ± 3%    ~    (p=0.222 n=9+9)
> PgbenchQuery_Cockroach-4            3.29ms ± 7%    3.22ms ± 5%    ~    (p=0.143 n=10+10)
> ParallelPgbenchQuery_Cockroach-4    3.08ms ± 5%    3.18ms ± 6%    ~    (p=0.079 n=9+10)
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/cockroachdb/cockroach/pull/4627
> Commit Summary
>
>    - util/interval: Remove Overlapper interface from IntervalTree
>    - util/interval: Make sure RangeGroup's key only escapes when needed
>
> File Changes
>
>    - *M* util/cache/cache.go
>    <https://github.com/cockroachdb/cockroach/pull/4627/files#diff-0> (16)
>    - *M* util/interval/interval.go
>    <https://github.com/cockroachdb/cockroach/pull/4627/files#diff-1> (86)
>    - *M* util/interval/range_group.go
>    <https://github.com/cockroachdb/cockroach/pull/4627/files#diff-2> (33)
>
> Patch Links:
>
>    - https://github.com/cockroachdb/cockroach/pull/4627.patch
>    - https://github.com/cockroachdb/cockroach/pull/4627.diff
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/4627>.
>