@nvanbenschoten 2016-02-25
This is actually a pretty substantial optimization in any situation where we have high contention for a part of the keyspace. I'll get more numbers on it, but what I've seen so far is that the `RangeQueue` reduced the average command dependency list by a little over **50x**, but it also became the **fourth** largest source of allocations (granted these are short lived allocations, so it is still an improvement). The problem was that we were allocating every time we wanted to get a key's overlapping ranges in the `IntervalTree` because we needed to fit the `Overlapper` interface. With this change, we only need to allocate when we are actually adding to the interval tree. The same story goes for the `TimestampCache`, we needed to allocate during every call to `GetOverlaps`, even though we basically implemented the interface in an identical way everywhere.

This is also more of a preliminary change to set up a larger change to `TimestampCache` soon.