@dt 2016-02-25
The variance is also lower on most of those numbers. I'd be curious to with `-benchmem` stats too.
Also, while deleting 100 rows might be unusual, inserting or updating 100 certainly isn't (in my experience) and 1% there is nice.