@nvanbenschoten 2016-02-25
To add to the previous numbers I gave after doing some quick profiling while running 3 instances of the Photos app with 50 users each:

Before this change:
- `interval.(*rangeTree).Add` did not show up when profiling for inuse_space
- However, when profiling for alloc_space, it was the **number 1** source of total allocations because of all of the short lived allocations
- The function allocated about twice as much space as `parser.(*sqlParserImpl).Parse` in total

After this change:
- `interval.(*rangeTree).Add` does not show up in any profiles...