@petermattis 2016-02-25
LGTM

---

Review status: 0 of 3 files reviewed at latest revision, 3 unresolved discussions.

---

*[util/interval/interval.go, line 61 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4627#-KBLJ3-mDKKaRfm_sfB1:-KBLJ3-nZKIrwHvrhrZk:-347708092) ([raw file](https://github.com/cockroachdb/cockroach/blob/a1e72b6d31b5579eab9242f193f9e94fc9ee1c7f/util/interval/interval.go#L61)):*
`s/inclusive/exclusive/g`

---

*[util/interval/interval.go, line 657 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4627#-KBLKRWX9QNiAp0XUKlQ:-KBLKRWX9QNiAp0XUKlR:-1844811914) ([raw file](https://github.com/cockroachdb/cockroach/blob/a1e72b6d31b5579eab9242f193f9e94fc9ee1c7f/util/interval/interval.go#L657)):*
Looks like some of the comments need updating. `q.Overlap` no longer exists.

---

*[util/interval/interval.go, line 691 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4627#-KBLJe3HAPPzBb-axZdl:-KBLJe3HAPPzBb-axZdm:1539291744) ([raw file](https://github.com/cockroachdb/cockroach/blob/a1e72b6d31b5579eab9242f193f9e94fc9ee1c7f/util/interval/interval.go#L691)):*
Would it be simpler to replace the `InclusiveRanges` field with `OverlapperFunc func(Range, Range) bool`?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4627#-:-KBLKqSpm-u05l2lX9JJ:676834233)*
<!-- Sent from Reviewable.io -->