@tschottdorf 2016-02-25
I think what we ought to be looking at is the `-benchmem` diff. Mind posting those? I'm sure they'll look goood (yes, triple o).