@dt 2016-02-25
Also, if you think that the benchmark results might be different in the presence of contention, you might want to rebase too before re-running ParallelPgbenchQuery, since that one wasn't really accurate before #4629.

---

Review status: 0 of 3 files reviewed at latest revision, 3 unresolved discussions.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4627#-:-KBLO1JBbsrrhObhs_qZ:-1039729417)*
<!-- Sent from Reviewable.io -->