@nvanbenschoten 2016-02-25
Decreasing the number of accounts in the parallel Pgbench benchmark to 5 to force more probabilistic contention has a pretty big effect on these stats.

```
name                              old time/op    new time/op    delta
ParallelPgbenchQuery_Cockroach-8    7.15ms ±17%    6.73ms ± 8%   -5.92%  (p=0.017 n=10+9)

name                              old alloc/op   new alloc/op   delta
ParallelPgbenchQuery_Cockroach-8     750kB ±23%     669kB ±12%  -10.73%  (p=0.017 n=10+9)

name                              old allocs/op  new allocs/op  delta
ParallelPgbenchQuery_Cockroach-8     9.48k ±12%     8.33k ±11%  -12.12%   (p=0.001 n=9+8)
```

Still, I don't think any of our benchmarks are even close to representative of a real world contentious scenario by just relying on `b.RunParallel`. This is something we should look into soon.